Added custom dimension validation/rationalisation
The limits for the dimensions of the custom mine field have been set to be between 1 and 30, exceeding either limit (so less than 1 or more than 30) forces the code to modify that dimension to the limit it has passed.  The limit for the number of mines is set to be the number of squares in the minefield minus one (i.e. there is always one square you can click on that is not a mine)