Ensure top-left square is never a mine
With the parameter rationalisation code in place it is safe to assume there will always be one square that is not a mine, I found the simplest (though obviously not the quickest) way to ensure the top-left square was clickable was to repeat the randomisation of the grid if there was a mine there.  I am sure I could find a better way of doing this by changing the "fisheryates" function but that felt slightly destructive.  On my mobile phone using a 30 x 30 grid with 899 mines (i.e. only one clickable square) it still takes only seconds to find that right combination.

To test this just keep clicking the "New game" button and clicking the top-left square.  Even on the "Beginner" setting it only took a few clicks to hit a mine before the change, after the change it did not happen.

This resolves #8