Replaced all ID references with class references
The code no longer references any element by its ID but instead uses
class names.  Where I have added a class to enable this I have prefixed
the name with "ms" to represent MineSweeper in an attempt to avoid
confusion with any classes that may be in use on the page in which this
object is embedded.